move test to GithubWebTests
